feat: add modal submit handler, move verifyrules into modal
