+++ src/interactions/modal/index.ts
@@ -1,0 +1,3 @@
+import { VerifyRulesModal } from './verifyRulesModal';
+
+export default [VerifyRulesModal];
